Handle connectivity in rollbacks
This required a small change to the connectivity calculation, using
a different valid-competitors metric that can be determined solely
from data available on match results, so that we can calculate it
without having to be in the middle of a full recalculation.

Rollback now recalculates a competitor's connectivity after deleting
events, and recalculates the project's connectivity baselines after
all rating change operations are done.